repotool test refactoring and cleanup.

+++ test/repotool-compare-tag-svn-git.sh
@@ -1,7 +1,4 @@
-## Test comparing tag between svn and git repository
-
-command -v svn >/dev/null 2>&1 || { echo "    Skipped, svn missing."; exit 0; }
-command -v git >/dev/null 2>&1 || { echo "    Skipped, git missing."; exit 0; }
+## Test comparing tag between svn and git repo
 
 # Results should be independent of what file stem this is, as
 # long as it's an svn dump and has the right festure to be named by cmploc.
@@ -10,6 +7,9 @@ cmploc=tag1
 cmpmode=-t
 
 # No user-serviceable parts below this line
+
+command -v svn >/dev/null 2>&1 || { echo "    Skipped, svn missing."; exit 0; }
+command -v git >/dev/null 2>&1 || { echo "    Skipped, git missing."; exit 0; }
 
 trap "rm -rf /tmp/test-repo$$-svn /tmp/test-repo$$-git /tmp/out$$" 0 12 2 15
